requirements.txt
# Flask App
Flask==0.10.1
wtforms==2.1
WTForms-JSON==0.2.10
SQLAlchemy==1.0.12
sqlalchemy-gevent==0.2
flask-sqlalchemy==2.1
graypy==0.2.9

# Gunicorn
gunicorn==19.4.5
gevent==1.1.1

# Testing
factory-boy==2.8.1
Flask-Testing==0.4.2
mock==1.3.0
